Refactor sysCallDispatcher and time module
This commit refactors the sysCallDispatcher module and the time module in the Kernel.

In sysCallDispatcher.c, the code changes remove a commented out line of code and update variable names for better readability. The changes improve the sys_write function by removing a call to the printChars function, which was previously used to print characters to the screen in red color. The code changes also update the sysCallDispatcher function to use more descriptive variable names.

In time.c, the code changes include adding a new header file, times.h, which contains function declarations for getting the current system time. The getTime function is modified to correctly format the system time string. The changes also update the variable names for better readability.

These code changes improve the overall organization and readability of the sysCallDispatcher and time modules in the Kernel.